% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sp.R
\name{add_sp}
\alias{add_sp}
\title{Add shapefiles on map of study area}
\usage{
add_sp(sp_path = NULL, col = NULL, border_col = NULL)
}
\arguments{
\item{sp_path}{Path of shapefiles.}

\item{col}{A color for added shapefile.}

\item{border_col}{A color for border of added shapefile.}
}
\value{

}
\description{
Add shapefiles on map of study area
}
\examples{
\dontrun{
add_sp(sp_path=".", col="blue", border_col="black")
}
}
