% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_points.R
\name{add_stations}
\alias{add_stations}
\title{Add sampling stations}
\usage{
add_stations(
  stations_x = NULL,
  stations_y = NULL,
  col = "red",
  points_cex = 0.5
)
}
\arguments{
\item{stations_x}{A vector of x coordinates of sampling stations}

\item{stations_y}{A vector of y coordiates of sampling stations}

\item{col}{Color of sampling stations}

\item{points_cex}{A numeric value giving the size of points}
}
\value{
A map with smapling stations
}
\description{
Add sampling stations
}
\examples{
\dontrun{
add_stations(station_x=c(-70.8, -67), station_y=c(42.5, 44.8), col="blue", point_ces=0.2)
}
}
