% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.data.dest.R
\docType{data}
\name{msa.data.dest}
\alias{msa.data.dest}
\title{msa.data.dest}
\format{A matrix of variable c(part, operator, reaction) of 30 rows}
\usage{
msa.dest
}
\description{
This is example dataset for msa.dest
}
\keyword{datasets}
