% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.data.gageRR.R
\docType{data}
\name{msa.data.gageRR}
\alias{msa.data.gageRR}
\title{msa.data.gageRR}
\format{A matrix of variable c("Part","Operator","Data") of 60 rows}
\usage{
msa.data.gageRR
}
\description{
This is example dataset for msa.gageRR
}
\keyword{datasets}
