% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.data.linBias.R
\docType{data}
\name{msa.data.linBias}
\alias{msa.data.linBias}
\title{msa.data.linBias}
\format{A matrix of variable c(part, standard, reaction) of 60 rows}
\usage{
msa.linBias
}
\description{
This is example dataset for msa.linBias
}
\keyword{datasets}
