% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.data.oneway.R
\docType{data}
\name{msa.data.oneway}
\alias{msa.data.oneway}
\title{msa.data.type3}
\format{A matrix of variable c("Part","Data") of 50 rows}
\usage{
msa.data.oneway
}
\description{
This is example dataset for msa.oneway
}
\keyword{datasets}
