% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.data.repBias.R
\docType{data}
\name{msa.data.repBias}
\alias{msa.data.repBias}
\title{msa.data.repBias}
\format{A matrix of variable c("diameter") of 50 rows}
\usage{
msa.repBias
}
\description{
This is example dataset for msa.repBias
}
\keyword{datasets}
