% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.data.repBias2.R
\docType{data}
\name{msa.data.repBias2}
\alias{msa.data.repBias2}
\title{msa.data.repBias2}
\format{A matrix of variable c("Data") of 50 rows}
\usage{
msa.repBias
}
\description{
This is example dataset for msa.repBias
}
\keyword{datasets}
