% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.destGageTable.R
\name{msa.destGageTable}
\alias{msa.destGageTable}
\title{MSA distructive Testing}
\usage{
msa.destGageTable(gdo, method = "nested", sigma = 6, alpha = 0.05,
  tolerance, dig = 5, main = "Gage R&R table", sub = "My MSA project")
}
\value{
msa.distTest
}
\description{
MSA distructive Testing
}
