% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.destTest.R
\name{msa.destTest}
\alias{msa.destTest}
\title{MSA destructive Test}
\usage{
msa.destTest(operators, parts, results, data, method = "nested", lsl, usl,
  alpha = 0.05, sigma = 6, digits = 5,
  main = "Gage R&R destructive Test", sub = "My MSA project")
}
\value{
msa.destTest
}
\description{
MSA destructive Test
}
