% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.gageRR.R
\name{msa.gageRR}
\alias{msa.gageRR}
\title{MSA gage R&R test}
\usage{
msa.gageRR(var, part, appr, lsl = NA, usl = NA, sigma = 6, data,
  main = "Gage R&R Study", sub = "My MSA project", alphaLim = 0.05,
  digits = 4)
}
\value{
msa.gageRR
}
\description{
MSA gage R&R test
}
