% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.gageRR2.R
\name{msa.gageRR2}
\alias{msa.gageRR2}
\title{MSA Type1 study analysis}
\usage{
msa.gageRR2(var, part, appr, lsl = NA, usl = NA, sigma = 6, data,
  main = "Gage R&R Study(ANOVA)", sub = "My MSA project", alphaLim = 0.05,
  digits = 5, errorTerm = "interaction")
}
\value{
msa.gageRR2
}
\description{
MSA Type1 study analysis
}
