% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.linBias.R
\name{msa.linBias}
\alias{msa.linBias}
\title{MSA linearity and bias test}
\usage{
msa.linBias(part, reference, output, data, conf.level = 0.95, ylim,
  main = "Gage Linearity and Bias Study", sub = "My MSA project")
}
\value{
msa.linBias
}
\description{
MSA linearity and bias test
}
