% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.oneway.R
\name{msa.oneway}
\alias{msa.oneway}
\title{MSA gageR&R oneway study analysis}
\usage{
msa.oneway(part, result, lsl, usl, data, alpha = 0.05, sigma = 6,
  digits = 5, main = "Gage Oneway Test", sub = "My MSA project")
}
\value{
msa.oneway
}
\description{
MSA gageR&R oneway study analysis
}
