% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundation.R
\name{msa.prepCanvas}
\alias{msa.prepCanvas}
\title{This is used to get prepare visualization}
\usage{
msa.prepCanvas(main = "Measurment System Analysis Graph",
  sub = "My MSA project", pca.col = c("#666666", "#BBBBBB", "#CCCCCC",
  "#DDDDDD", "#EEEEEE"))
}
\value{
prep Canvas
}
\description{
This is used to get prepare visualization
}
