% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.repBias.R
\name{msa.repBias}
\alias{msa.repBias}
\title{MSA Reapeatability and Bias test}
\usage{
msa.repBias(data, tol, LSL, USL, resol, target, alpha = 0.05,
  mainTitle = "Gage Reapeatability and Bias Test",
  subTitle = "My MSA project")
}
\value{
msa.repBias
}
\description{
MSA Reapeatability and Bias test
}
