% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.type1.R
\name{msa.type1}
\alias{msa.type1}
\title{MSA Type1 study analysis}
\usage{
msa.type1(data, tol, LSL, USL, resol, target, alpha = 0.05,
  mainTitle = "Measurment System Analysis Graph",
  subTitle = "My MSA project")
}
\value{
type1check
}
\description{
MSA Type1 study analysis
}
