% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{MSA_feedback_with_graph}
\alias{MSA_feedback_with_graph}
\title{MSA feedback (with graph)}
\usage{
MSA_feedback_with_graph(dict = MSA::MSA_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual and graphical feedback at the end of the test.
}
\examples{
\dontrun{
MSA_demo(feedback = MSA_feedback_with_score())}
}
