% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSA_itembank_training_only}
\alias{MSA_itembank_training_only}
\title{MSA_itembank_training_only}
\format{
A data frame with X rows and Y columns:
}
\source{
MSA_itembank_training_only
}
\usage{
MSA_itembank_training_only
}
\description{
MSA_itembank_training_only

A brief description of the MSA_itembank_training_only dataset.
This dataset is intended for internal use within the package.
}
\keyword{datasets}
