% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MSA_standalone}
\alias{MSA_standalone}
\title{Standalone MSA}
\usage{
MSA_standalone(
  title = NULL,
  num_items = 18L,
  with_id = TRUE,
  with_feedback = TRUE,
  with_welcome = TRUE,
  take_training = FALSE,
  admin_password = "password",
  researcher_email = "put.your.email-adress@here",
  languages = c("en", "de"),
  dict = MSA::MSA_dict,
  validate_id = "auto",
  balance_over = c("target_instrument", "complexity", "level"),
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{num_items}{(Scalar integer) Number of items to be adminstered.}

\item{with_id}{(Scalar boolean) Indicates, if ID should be asked for. Defaults to TRUE}

\item{with_feedback}{(Scalar boolean) Indicates if performance feedback will be given at the end of the test. Defaults to  FALSE}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed. Defaults to TRUE}

\item{take_training}{(Scalar boolean) Enable practice session before the actual session. Defaults to FALSE}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"en"}) and German (\code{"de"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto" for default validation
which means ID should consist only of  alphanumeric characters.}

\item{balance_over}{(Character vector) Indicates how items are selected from the item pool. Balance means that the proportion of items for each parameter is equal.
Possible parameters are
"target_instrument": the target instrument; balancing = equal proportion of the four different instruments ('Lead Voice', 'Piano', 'Guitar', 'Bass').
"complexity": the musical complexity, i.e. number of instruments within the mixture; balancing = equal proportion of items with 'three' and 'six' instruments playing in the mixture.
"level": the level-ratio between target and the mixture; balancing = equal proportion of items with '0', '-5', '-10', '-15' level-ratios.
Default is a fully balanced design: c("target_instrument", "complexity", "level").
Note: By default, there is always an equal proportion of "with target instrument" and "without target" items in the pool.}

\item{...}{Further arguments to be passed to \code{\link{MSA}()}.}
}
\description{
This function launches a standalone testing session for the MSA.
This can be used for data collection, either in the laboratory or online.
}
