% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_grid.R
\name{plot_map_grid}
\alias{plot_map_grid}
\title{Plot map of study area and sampling grids}
\usage{
plot_map_grid(
  map_sp_path = NULL,
  grids_sp_path = NULL,
  grids_col_var = NULL,
  grids_col = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = "Longitude",
  ylab = "Latitude",
  map_col = "gray95",
  map_border_col = "black",
  grids_border_col = "black",
  legend_x = NULL,
  legend_y = NULL,
  legend = NULL,
  legend_title = NULL
)
}
\arguments{
\item{map_sp_path}{Path of map shapefiles.}

\item{grids_sp_path}{Path of sampling grids shapefiles.}

\item{grids_col_var}{A variable to be used to plot strata (e.g., depth or longitudinal regions) with different colors.}

\item{grids_col}{A color palette to be used to plot strata (e.g., depth or longitudinal regions).}

\item{xlim}{Numeric vectors of length 2, giving the longitude range.}

\item{ylim}{Numeric vectors of length 2, giving the latitude range.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{map_col}{A color for land.}

\item{map_border_col}{A color for border of land.}

\item{grids_border_col}{A color for border of grids.}

\item{legend_x}{X coordinate to be used to position the legend.}

\item{legend_y}{Y coordinate to be used to position the legend.}

\item{legend}{A character vector to appear in the legend}

\item{legend_title}{A character string to appear at the top of the legend.}
}
\value{
A map of study area and sampling grids
}
\description{
Plot map of study area and sampling grids
}
\examples{
\dontrun{
plot_map_grid(xlim=c(-70.8, -67), ylim=c(42.5, 44.8))
}
}
