% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA.R
\name{MSA}
\alias{MSA}
\title{MSA}
\usage{
MSA(
  num_items = 18L,
  with_welcome = TRUE,
  take_training = FALSE,
  with_finish = TRUE,
  unique_songs_only = FALSE,
  label = "MSA_results",
  feedback = MSA::MSA_feedback_with_score(),
  dict = MSA::MSA_dict,
  with_target_in_mix = "balanced",
  target_instrument = "random",
  complexity = "random",
  level = "random"
)
}
\arguments{
\item{num_items}{(Scalar integer) Number of items to be administered.}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed. Defaults to TRUE}

\item{take_training}{(Scalar boolean) Enable practice session before the actual session. Defaults to FALSE}

\item{with_finish}{(Scalar boolean) Indicates, if a finish (not final!) page shall be displayed. Defaults to TRUE}

\item{unique_songs_only}{(Scalar boolean) Set to 'TRUE' if each song (base track) should play once in the dataset.
Please be aware, that the item pool only consists of 98 different base tracks. Defaults to FALSE}

\item{label}{(Scalar character) Label to give the MSA results in the output file.}

\item{feedback}{(Function) Defines the feedback to give the participant at the end of the test.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{with_target_in_mix}{(Character scalar) With target in the mixture (with_target_in_mix): Indicates how items are selected from the item pool.
Possible values are ("balanced") = equal proportion of items 'with target' and 'without target' in the mixture are in the item pool;
and ("random") = pick items randomly for this variable; Default to "balanced".}

\item{target_instrument}{(Character scalar) Target Instruments (TargetIns): Indicates how items are selected from the item pool.
Possible values are ("balanced") = equal proportion of the four different instruments ('Lead Voice', 'Piano', 'Guitar', 'Bass')
are selected for the item pool;
and ("random") = pick items randomly for this variable; Default to "random"}

\item{complexity}{(Character scalar) Musical complexity (i.e., Number of instruments within the mixture):
Indicates how items are selected from the item pool.
Possible values are ("balanced") = equal proportion of items with 'three' and 'six' instruments playing in the mixture
are selected for the item pool,
and ("random") =  pick items randomly for this variable; Default to "random"}

\item{level}{(Character scalar) Level-ratio between target and the mixture (level): Indicates how items are selected from the item pool.
Possible values are ("balanced") = equal proportion of items with '0', '-5', '-10', '-15' level-ratios are selected for the item pool,
and ("random") =  pick items randomly for this variable; Default to "random"}
}
\description{
This function defines a MSA  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MSA in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
}
\details{
For demoing the MSA, consider using \code{\link{MSA_demo}()}.
For a standalone implementation of the MSA,
consider using \code{\link{MSA_standalone}()}.
This can be used for data collection, either in the laboratory or online.
}
