% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA.R
\name{MSA}
\alias{MSA}
\title{MSA}
\usage{
MSA(
  num_items = 18L,
  with_welcome = TRUE,
  take_training = FALSE,
  with_finish = TRUE,
  label = "MSA_results",
  with_feedback = TRUE,
  feedback = MSA::MSA_feedback_with_score(),
  dict = MSA::MSA_dict,
  balance_over = c("target_instrument", "complexity", "level")
)
}
\arguments{
\item{num_items}{(Scalar integer) Number of items to be administered.}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed. Defaults to TRUE}

\item{take_training}{(Scalar boolean) Enable practice session before the actual session. Defaults to FALSE}

\item{with_finish}{(Scalar boolean) Indicates, if a finish (not final!) page shall be displayed. Defaults to TRUE}

\item{label}{(Scalar character) Label to give the MSA results in the output file.}

\item{with_feedback}{(Scalar boolean) Defines whether the test person receives feedback at the end of the task.}

\item{feedback}{(Function) Defines the feedback to give the participant at the end of the test.}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{balance_over}{(Character vector) Indicates how items are selected from the item pool. Balance means that the proportion of items for each parameter is equal.
Possible parameters are
"target_instrument": the target instrument; balancing = equal proportion of the four different instruments ('Lead Voice', 'Piano', 'Guitar', 'Bass').
"complexity": the musical complexity, i.e. number of instruments within the mixture; balancing = equal proportion of items with 'three' and 'six' instruments.
"level": the level-ratio between target and the mixture; balancing = equal proportion of items with '0', '-5', '-10', '-15' level-ratios.
Default is a fully balanced design: c("target_instrument", "complexity", "level").
Note: By default, there is always an equal proportion of "with target instrument" and "without target" items in the pool.}
}
\description{
This function defines a MSA  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MSA in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
The MSA is an non-adaptive instrument to test musical scene analysis
abilities among very diverse populations. An adaptive version of this
instrument is currently in development (expected early 2022).
The test is based on a two-alternative-forced-choice task - for a more
detailed description take a look into the README file.
}
\details{
For demoing the MSA, consider using \code{\link{MSA_demo}()}.
For a standalone implementation of the MSA,
consider using \code{\link{MSA_standalone}()}.
This can be used for data collection, either in the laboratory or online.
}
