% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSA.R
\name{MSA}
\alias{MSA}
\title{MSA}
\usage{
MSA(
  label = "MSA_results",
  num_items = 30L,
  with_welcome = TRUE,
  with_finish = TRUE,
  take_training = TRUE,
  with_video = FALSE,
  with_picture = FALSE,
  with_feedback = TRUE,
  feedback = MSA::MSA_feedback_with_graph(),
  balance_over = c("target_instrument", "complexity", "level"),
  dict = MSA::MSA_dict,
  adaptive = TRUE,
  long_version = FALSE,
  location_stim = "https://media.gold-msi.org/test_materials/MSAT",
  next_item.criterion = "bOpt",
  next_item.estimator = "BM",
  next_item.prior_dist = "norm",
  next_item.prior_par = c(0, 1),
  final_ability.estimator = "WL",
  constrain_answers = FALSE
)
}
\arguments{
\item{label}{(Scalar character) Label to give the MSA results in the output file.}

\item{num_items}{(Scalar integer) Number of items to be administered. We recommend to use at least 30 items (default)
for the adaptive MSA and at least 40 for the non-adaptive version.}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed. Defaults to TRUE}

\item{with_finish}{(Scalar boolean) Indicates, if a finish, i.e. a completion (not final!) page shall be displayed. Defaults to TRUE}

\item{take_training}{(Scalar boolean) Enable practice session before the actual session. Defaults to TRUE}

\item{with_video}{(Scalar boolean) Includes a demonstrative video in addition to the default text description,
to be used as an explanation for the practice session (take_training needs to be TRUE). Defaults to FALSE
Note: Correct display depends on the playback device. We recommend the use of Google Chrome.}

\item{with_picture}{(Scalar boolean) Includes a demonstrative picture in addition to the default text description,
to be used as an explanation for the practice session (take_training needs to be TRUE). Defaults to FALSE
Note: Correct display depends on the playback device. We recommend the use of Google Chrome.}

\item{with_feedback}{(Scalar boolean) Defines whether the test person receives feedback at the end of the task.}

\item{feedback}{(Function) Defines the feedback to give the participant at the end of the test.
Options are "MSA::MSA_feedback_with_graph()" (Default) and "MSA::MSA_feedback_with_score()" for showing only the number of correctly detected items.}

\item{balance_over}{(Character vector) Indicates how items are selected from the item pool. Balance means that the proportion of items for each parameter is equal.
Please note that this option is only available for the non-adaptive Version of the MSA (adaptive = FALSE).
"target_instrument": the target instrument; balancing = equal proportion of the four different instruments ('Lead Voice', 'Piano', 'Guitar', 'Bass').
"complexity": the musical complexity, i.e. number of instruments within the mixture; balancing = equal proportion of items with 'three' and 'six' instruments.
"level": the level-ratio between target and the mixture; balancing = equal proportion of items with '0', '-5', '-10', '-15' level-ratios.
Default is a fully balanced design: c("target_instrument", "complexity", "level").
Note: By default, there is always an equal proportion of "with target instrument" and "without target" items in the pool.}

\item{dict}{The psychTestR (package) dictionary used for internationalisation.}

\item{adaptive}{(Scalar boolean) Indicates whether you want to use the adaptive MSA (TRUE)
or the non-adaptive MSA (FASLE). Default is adaptive = TRUE.}

\item{long_version}{(Scalar boolean) Indicates whether you want to use the musical excerpts that include 8 seconds of music prior to the task (TRUE):
If (TRUE), each trail has the following procedure:
Part 1: the initial music segment (8s);
Part 2: after 1s pause, a single instrument (lead voice, bass, guitar, or piano) plays;
Part 3: after 1s pause, a mix of instruments plays, possibly including the target instrument.
If (FALSE), Part 1 is skipped (the original version; Default).}

\item{location_stim}{(Scalar character) Specify the location of the stimuli.
Default is the location on the DOTS servers from the "Deutsche Gesellschaft für Musikpsychologie "https://media.gold-msi.org/test_materials/MSAT".
To start the test locally (without internet connection), one have to download all stimuli from the provided drives:
Stimuli long version (https://drive.google.com/drive/folders/1OI2Ii2C8yUGu8M9BzyV3HVEiKv1YFXLf?usp=sharing)
Stimuli short version (https://drive.google.com/drive/folders/1cxPEOyAaipXFtWNEBaBOyFkImftlEOSz?usp=sharing)
Videos, which should belong into the same folder as the stimuli (https://drive.google.com/drive/folders/1_za6FNNnXWREk6NRAewvXhQsn9IG3TY8?usp=sharing)
Then, create a local host server using e.g.,"servr" package (servr::httd("C://Users//PC NAME//Stimuli folder path")) and then provide the
new URL, e.g., location_stim = "http://127.0.0.1:4321".}

\item{next_item.criterion}{(Character scalar)
Criterion for selecting successive items in the adaptive test.
See the \code{criterion} argument in \code{\link[catR]{nextItem}} for possible values.
Defaults to \code{"bOpt"}.}

\item{next_item.estimator}{(Character scalar)
Ability estimation method used for selecting successive items in the adaptive test.
See the \code{method} argument in \code{\link[catR]{thetaEst}} for possible values.
\code{"BM"}, Bayes modal,
corresponds to the setting used (will be further explained in the to be published MSA paper).
\code{"WL"}, weighted likelihood,
corresponds to the default setting used in versions <= 0.2.0 of this package.}

\item{next_item.prior_dist}{(Character scalar)
The type of prior distribution to use when calculating ability estimates
for item selection.
Ignored if \code{next_item.estimator} is not a Bayesian method.
Defaults to \code{"norm"} for a normal distribution.
See the \code{priorDist} argument in \code{\link[catR]{thetaEst}} for possible values.}

\item{next_item.prior_par}{(Numeric vector, length 2)
Parameters for the prior distribution;
see the \code{priorPar} argument in \code{\link[catR]{thetaEst}} for details.
Ignored if \code{next_item.estimator} is not a Bayesian method.
The dfeault is \code{c(0, 1)}.}

\item{final_ability.estimator}{Estimation method used for the final ability estimate.
See the \code{method} argument in \code{\link[catR]{thetaEst}} for possible values.
The default is \code{"WL"}, weighted likelihood.
#' If a Bayesian method is chosen, its prior distribution will be defined
by the \code{next_item.prior_dist} and \code{next_item.prior_par} arguments.}

\item{constrain_answers}{(Logical scalar)
If \code{TRUE}, then item selection will be constrained so that the
correct answers are distributed as evenly as possible over the course of the test.
We recommend leaving this option disabled.}
}
\description{
This function defines a MSA  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the MSA in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
The MSA is an adaptive instrument to test musical scene analysis
abilities among very diverse populations. An non-adaptive version is also available.
The test is based on a two-alternative-forced-choice task - for a more
detailed description take a look at the Readme file on GITHUB (https://github.com/rhake14/MSA/tree/main#readme)
}
\details{
For demoing the MSA, consider using \code{\link{MSA_demo}()}.
For a standalone implementation of the MSA,
consider using \code{\link{MSA_standalone}()}.
This can be used for data collection, either in the laboratory or online.
}
