% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{MSA_standalone}
\alias{MSA_standalone}
\title{Standalone MSA}
\usage{
MSA_standalone(
  title = NULL,
  num_items = 18L,
  with_id = TRUE,
  with_feedback = TRUE,
  with_welcome = TRUE,
  take_training = FALSE,
  unique_songs_only = FALSE,
  admin_password = "password",
  researcher_email = "put.your.email-adress@here",
  languages = c("en", "de"),
  dict = MSA::MSA_dict,
  validate_id = "auto",
  with_target_in_mix = "balanced",
  target_instrument = "random",
  complexity = "random",
  level = "random",
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{num_items}{(Scalar integer) Number of items to be adminstered.}

\item{with_id}{(Scalar boolean) Indicates, if ID should be asked for. Defaults to TRUE}

\item{with_feedback}{(Scalar boolean) Indicates if performance feedback will be given at the end of the test. Defaults to  FALSE}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed. Defaults to TRUE}

\item{take_training}{(Scalar boolean) Enable practice session before the actual session. Defaults to FALSE}

\item{unique_songs_only}{(Scalar boolean) Set to 'TRUE' if each song (base track) should play once in the dataset.
Please be aware, that the item pool only consists of 98 different base tracks. Defaults to FALSE}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"en"}) and German (\code{"de"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{validate_id}{(Character scalar or closure) Function for validating IDs or string "auto" for default validation
which means ID should consist only of  alphanumeric characters.}

\item{with_target_in_mix}{(Character scalar) With target in the mixture (with_target_in_mix): Indicates how items are selected from the item pool.
Possible values are ("balanced") = equal proportion of items 'with target' and 'without target' in the mixture are in the item pool;
and ("random") = pick items randomly for this variable; Default to "balanced".}

\item{target_instrument}{(Character scalar) Target Instruments (TargetIns): Indicates how items are selected from the item pool.
Possible values are ("balanced") = equal proportion of the four different instruments ('Lead Voice', 'Piano', 'Guitar', 'Bass')
are selected for the item pool;
and ("random") = pick items randomly for this variable; Default to "random"}

\item{complexity}{(Character scalar) Musical complexity (i.e., Number of instruments within the mixture):
Indicates how items are selected from the item pool.
Possible values are ("balanced") = equal proportion of items with 'three' and 'six' instruments playing in the mixture
are selected for the item pool,
and ("random") =  pick items randomly for this variable; Default to "random"}

\item{level}{(Character scalar) Level-ratio between target and the mixture (level): Indicates how items are selected from the item pool.
Possible values are ("balanced") = equal proportion of items with '0', '-5', '-10', '-15' level-ratios are selected for the item pool,
and ("random") =  pick items randomly for this variable; Default to "random"}

\item{...}{Further arguments to be passed to \code{\link{MSA}()}.}
}
\description{
This function launches a standalone testing session for the MSA.
This can be used for data collection, either in the laboratory or online.
}
